Move CompactionTaskStatusStore to core
