+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/task/CompactionTask.java
@@ -40,6 +40,7 @@ import sleeper.core.table.TableNotFoundException;
 import sleeper.core.tracker.compaction.job.CompactionJobTracker;
 import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
 import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatusStore;
 import sleeper.core.util.LoggedDuration;
 
 import java.io.IOException;
