+++ java/compaction/compaction-status-store/src/test/java/sleeper/compaction/status/store/task/StoreCompactionTaskExpiryIT.java
@@ -17,9 +17,9 @@ package sleeper.compaction.status.store.task;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.compaction.core.task.CompactionTaskStatusStore;
 import sleeper.compaction.status.store.testutils.DynamoDBCompactionTaskStatusStoreTestBase;
 import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatusStore;
 
 import java.time.Duration;
 import java.time.Instant;
